scripts/requirements.txt
psycopg2-binary >= 2.8.6
PyMySQL >= 1.0.2